Part a

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -27,6 +27,9 @@ public class ImageController {
     @Autowired
     private TagService tagService;
 
+    public static final String EDIT_IMAGE_ERROR = "Only the owner of the image can edit the image";
+    public static final String DELETE_IMAGE_ERROR = "Only the owner of the image can delete the image";
+
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
     public String getUserImages(Model model) {
@@ -45,9 +48,10 @@ public class ImageController {
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
-    @RequestMapping("/images/{title}")
-    public String showImage(@PathVariable("title") String title, Model model) {
-        Image image = imageService.getImageByTitle(title);
+    @RequestMapping("/images/{imageId}/{title}")
+    public String showImage(@PathVariable("imageId") Integer imageId, @PathVariable("title") String title, Model model) {
+        //Image image = imageService.getImageByTitle(title);
+        Image image = imageService.getImage(imageId);
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
         return "images/image";
@@ -92,12 +96,18 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
 
-        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
+
+        if(image.getTags() != null && !image.getTags().isEmpty())
+            model.addAttribute("tags", convertTagsToString(image.getTags()));
+
+        if(!isUserLoggedIn(image, session)){
+            model.addAttribute("editError", EDIT_IMAGE_ERROR);
+            return "images/image";
+        }
         return "images/edit";
     }
 
@@ -140,11 +150,46 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        Image image = imageService.getImage(imageId);
+        if(!isUserLoggedIn(image, session)){
+            model.addAttribute("deleteError", DELETE_IMAGE_ERROR);
+            model.addAttribute("image", image);
+            if(image.getTags() != null && !image.getTags().isEmpty())
+                model.addAttribute("tags",convertTagsToString(image.getTags()));
+            return "images/image";
+        }
+
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
 
+
+    //This method is used for getting the logged in user details
+    //returns null if the loggeduser object is empty
+    //else returns the loggeduser info
+    private User getLoggedInUser(HttpSession session) {
+        Object loggedUser = session.getAttribute("loggeduser");
+        if (loggedUser == null) {
+            return null;
+        }
+        return (User) loggedUser;
+    }
+
+    //This method is used to check whether the user has logged in
+    private boolean isUserLoggedIn(Image image, HttpSession session){
+        User loginUser = getLoggedInUser(session);
+        if(image == null || loginUser == null){
+            return false;
+        }
+        if(loginUser.getUsername() != null
+            && image.getUser() != null &&
+                image.getUser().getUsername() != null &&
+                    image.getUser().getUsername().equals(loginUser.getUsername())){
+            return true;
+        }
+        return false;
+    }
 
     //This method converts the image to Base64 format
     private String convertUploadedFileToBase64(MultipartFile file) throws IOException {
